backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
pydantic-settings
httpx
python-multipart
websockets
aiofiles
